+++ src/parents/Moving_GravityView.java
@@ -16,7 +16,7 @@ public class Moving_GravityView extends MovingView implements Gravity{
 	private int gravityThreshold;
 	private boolean atThreshold;
 	
-	public Moving_GravityView(Context context,double movingSpeedY,double movingSpeedX,int width,int height,int imageId){
+	public Moving_GravityView(Context context,float movingSpeedY,float movingSpeedX,int width,int height,int imageId){
 		super(context, movingSpeedY, movingSpeedX, width, height, imageId);
 
 		atThreshold=false;
